self.hdr wasnt float

+++ timelapse.py
@@ -199,7 +199,7 @@ class timelapse:
         """
         Take an additional under- and overexposed image.
         """
-        compensation = self.currentss * (float(self.hdr)/100)
+        compensation = self.currentss * (self.hdr/100)
         print self.currentss, self.hdr/100
         self.camera.shutter_speed = self.currentss - compensation
         print self.camera.shutter_speed
@@ -342,7 +342,7 @@ def main(argv):
     parser.add_argument('-c', '--colourbalance', nargs=2, default='auto',
                         type=str, help='Set white balance as red and blue. '
                                        '''\nEg. \'493/256\' '387/256\' ''')
-    parser.add_argument('-r', '--hdr', default=0, type=int,
+    parser.add_argument('-r', '--hdr', default=0, type=float,
                         help='Take two additional images, one under-, one '
                                'overexposed. \n Set this from 1 to 25,'
                                'depending the desired difference in exposure')
